# api.yaml
openapi: 3.0.1
info:
  title: Bchat Social Group Server
  description: >
    API documentation for Bchat Social groups.  This is the API that
    [Bchat](https://bchat.beldex.io/) and related tools use to interface with Social groups.
  contact:
    name: The Beldex Project
    email: support@beldex.io
    url: https://bchat.beldex.io/
  license:
    name: GPL v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: "3.0"
...
paths:
...
  /beldex/v4/lsrpc:
    post:
      tags: [Onion]
      summary: "Endpoint for submitting an encrypted onion request"
      description: >
        The data is an onion-encrypted request for another endpoint, wrapped in an encrypted layer
        for this Social group server.
      requestBody:
        description: >
          Onion request endpoint information.  Note that the field descriptions here are *inside*
          the onion encryption layer; the onion encryption payload itself is documented elsewhere.
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: >
                The onion request data.  This is encoded/encrypted in multiple layers, as follows.
                
                
                The data is first constructed as one or two parts: the first part is json contains
                request metadata with fields:
                - method -- "GET", "POST", etc. of the subrequest
                - endpoint -- the subrequest endpoint, e.g. `/room/some-room/messages/since/123`
                - headers -- request headers, typically containing X-BSGS-* auth info and, for
                  POST/PUT requests, a Content-Type.
                
                
                The second part is the request body, in bytes (only for POST/PUT requests).
                
                
                These two pieces are encoded as a one- or two-string bencoded list, which has format:
                
                
                `l123:jsone` or `l123:json456:bodye` where 123 is the length of the json and 456 is
                the length of the body, if the request has a body.  (Both strings are byte strings).
                
                
                This data is then encrypted using onion-request style encryption; see the
                beldex-storage-server for details on how this is done.
      responses:
        200:
          description: >
            Onion request completed.  The subrequest response will be encoded in a two-string
            bencoded list (see the request details for the encoding specifics) where the first
            string contains the response metadata as json with keys:
            
            
            - code -- the HTTP response code of the subrequest, e.g. 200, 404
            - headers -- a dict of HTTP response headers; the header name keys are always
              lower-cased.
            
            
            The second part is the response body bytes; as in HTTP, interpreting this depends on the
            `content-type` header in the `headers` metadata, and the details of the invoked
            endpoint.
            
            
            These two byte strings are bencoded, and then the bencoded data is encrypted using the
            same ephemeral key and encryption type as was used in the request.
components:
...
  securitySchemes:
    pubkey:
      type: apiKey
      name: X-BSGS-Pubkey
      in: header
      description: >
        The Ed25519 public key of the request.  For non-blinded requests this is the root bchat
        Ed25519 pubkey with '00' prefixed; for blinded requests this begins with 'bd' and follows
        the blinding procedure detailed elsewhere.
    nonce:
      type: apiKey
      name: X-BSGS-Nonce
      in: header
      description: >
        A unique, random nonce string of exactly 16 source bytes encoded in base64 (i.e. 24 base64
        characters, including two trailing padding characters).  This must be unique for every
        request from this pubkey within the last 24 hours; nonce reuse will result in failed
        requests.
    timestamp:
      type: apiKey
      name: X-BSGS-Timestamp
      in: header
      description: >
        Unix timestamp integer (expressed as a string) of the time when the request was initiated to
        help avoid replay attacks.  This timestamp must be within ±24 hours of the server's time
        when the request is received.
    signature:
      type: apiKey
      name: X-BSGS-Signature
      in: header
      description: >
        Ed25519 signature of
        
        
        `SERVER_PUBKEY || NONCE || TIMESTAMP || METHOD || PATH || HBODY`
        
        
        signed using the client's blinded or unblinded pubkey (from the `X-BSGS-Pubkey` header),
        encoded using base64 (with or without padding).
        
        
...